% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_network.R
\name{exampleNetwork}
\alias{exampleNetwork}
\title{Example Network}
\usage{
exampleNetwork(n_links = 30L, index = 1L)
}
\arguments{
\item{n_links}{number of links that the network to be returned shall contain. 
If there is no network with the given number of links, a network that is 
slightly smaller or bigger is returned. Giving -1L here returns the biggest
possible network that is stored in this package.}

\item{index}{integer number to switch between different sub nets if more 
than one subset of the required size is available.}
}
\value{
data frame with (roughly) \code{n_links} observations of two
  variables. The variables \code{us_node_id} (upstream node ID) and
  \code{ds_node_id} (downstream node ID) define the connections between
  links. They are needed if the list of connected links upstream of each node
  is to be calculated by means of \code{\link{getConnectedLinks}}.
}
\description{
Example data describing a network of connected links
}
