% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connected_links.R
\name{getConnectedLinks.C}
\alias{getConnectedLinks.C}
\title{Get Connected Links (C Implementation)}
\usage{
getConnectedLinks.C(
  directLinks,
  resultSize = 60000,
  queueSize = 1024 * 1024,
  version = 1,
  dbg = FALSE
)
}
\arguments{
\item{directLinks}{directLinks}

\item{resultSize}{default: 60000}

\item{queueSize}{default: 1024*1024}

\item{version}{version of C implementation: 1,2 or 3 (default: 1)}

\item{dbg}{default: FALSE}
}
\value{
data.frame with columns \code{origin}, \code{linked} and attribute
  \code{numberOfConnected}
}
\description{
Get Connected Links (C Implementation)
}
