% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphGetConnected.R
\name{getConnectedLinks.R}
\alias{getConnectedLinks.R}
\title{Get Connected Links (R Implementation)}
\usage{
getConnectedLinks.R(directly.connected, dbg = FALSE)
}
\arguments{
\item{directly.connected}{directly.connected}

\item{dbg}{default: FALSE}
}
\value{
get connected links with R implementation
}
\description{
Get Connected Links (R Implementation)
}
