% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphGetConnected.R
\name{getConnectedLinks}
\alias{getConnectedLinks}
\title{Get Connected Links}
\usage{
getConnectedLinks(x, upstream = TRUE, version = 1, dbg = FALSE, ...)
}
\arguments{
\item{x}{data frame with each row representing a link of the network. Required
columns: \emph{us_node_id}, \emph{ds_node_id}}

\item{upstream}{if TRUE (upstream), if FALSE (downstream), default: TRUE}

\item{version}{1: R-implementation, 2: C-implementation, (default: 1)}

\item{dbg}{default: FALSE}

\item{...}{additional arguments passed to  \code{\link{getConnectedLinks.C}}}
}
\value{
get connected links
}
\description{
Get Connected Links
}
