% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphGetConnected.R
\name{getDirectLinks.C}
\alias{getDirectLinks.C}
\title{Get Direct Links (C Implementation)}
\usage{
getDirectLinks.C(x, MAX_DIRECT_CONNECTIONS = 5, dbg = FALSE)
}
\arguments{
\item{x}{data frame with each row representing a link of the network. Required
columns: \emph{us_node_id}, \emph{ds_node_id}}

\item{MAX_DIRECT_CONNECTIONS}{default: 5}

\item{dbg}{default: FALSE}
}
\value{
get direct links with C implementation
}
\description{
Get Direct Links (C Implementation)
}
