% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphGetConnected.R
\name{exampleNetwork}
\alias{exampleNetwork}
\title{Example network}
\usage{
exampleNetwork()
}
\value{
data frame with 10040 observations of two variables. The variables
  \code{us_node_id} (upstream node ID) and \code{ds_node_id} (downstream node
  ID) define the connections between links. They are needed if the list of
  connected links upstream of each node is to be calculated by means of
  \code{\link{getConnectedLinks}}.
}
\description{
Example data describing a network of connected links
}
