% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphGetConnected.R
\name{getConnectedLinks.C}
\alias{getConnectedLinks.C}
\title{#' Get Connected Links (C Implementation)}
\usage{
getConnectedLinks.C(
  directLinks,
  resultSize = 60000,
  queueSize = 1024 * 1024,
  version = 1,
  dbg = FALSE
)
}
\arguments{
\item{directLinks}{directLinks}

\item{resultSize}{default: 60000}

\item{queueSize}{default: 1024*1024}

\item{version}{version of C implementation: 1,2 or 3 (default: 1)}

\item{dbg}{default: FALSE}
}
\value{
get connected links with C implementation
}
\description{
#' Get Connected Links (C Implementation)
}
