% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connected_links.R
\name{getConnectedLinks.R}
\alias{getConnectedLinks.R}
\title{Get Connected Links (R Implementation)}
\usage{
getConnectedLinks.R(directLinks, dbg = FALSE)
}
\arguments{
\item{directLinks}{direct links}

\item{dbg}{default: FALSE}
}
\value{
get connected links with R implementation
}
\description{
Get Connected Links (R Implementation)
}
