% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphGetConnected.R
\name{getConnectedLinks}
\alias{getConnectedLinks}
\title{Get Connected Links}
\usage{
getConnectedLinks(x, upstream = TRUE, version = 1, dbg = FALSE, ...)
}
\arguments{
\item{x}{data frame with each row representing a link of the network. Required
columns: \emph{us_node_id}, \emph{ds_node_id}}

\item{upstream}{if TRUE (upstream), if FALSE (downstream), default: TRUE}

\item{version}{1: R-implementation, 2: C-implementation, (default: 1)}

\item{dbg}{default: FALSE}

\item{...}{additional arguments passed to  \code{\link{getConnectedLinks.C}}}
}
\value{
get connected links
}
\description{
Get Connected Links
}
\examples{
network <- kwb.graph::exampleNetwork(n_links = -1L)

runtime.R <- vector()
runtime.C1 <- vector()  
runtime.C2 <- vector()  

n <- 1
resultSize <- 2054851
queueSize <- 100*1024

elapsed <- function(exp) system.time(exp)["elapsed"]

for (i in 1:n) {
  cat("run", i, "/", n, "\n")
  runtime.R[i] <- elapsed(x1 <- getConnectedLinks(
    network
  ))
  runtime.C1[i] <- elapsed(x2 <- getConnectedLinks(
    network, version = 2, resultSize = resultSize, queueSize = queueSize
  ))
  runtime.C2[i] <- elapsed(x3 <- getConnectedLinks(
    network, version = 3, resultSize = resultSize, queueSize = queueSize
  ))
}

cat("mean runtime with R-functions:", mean(runtime.R), "\n")
cat("mean runtime with C-functions(1):", mean(runtime.C1), "\n")  
cat("mean runtime with C-functions(2):", mean(runtime.C2), "\n")  

runtimeData <- data.frame(
  version = 1:3, 
  runtime = c(runtime.R), runtime.C1, runtime.C2
)

boxplot(runtime ~ version, data = runtimeData)

}
