% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connected_links.R
\name{getDirectLinks.R}
\alias{getDirectLinks.R}
\title{Get Direct Links (R Implementation)}
\usage{
getDirectLinks.R(x, upstream = TRUE)
}
\arguments{
\item{x}{data frame with each row representing a link of the network. Required
columns: \emph{us_node_id}, \emph{ds_node_id}}

\item{upstream}{if TRUE (upstream), if FALSE (downstream), default: TRUE}
}
\value{
list with as many elements as there are rows in \code{x}
}
\description{
Get Direct Links (R Implementation)
}
